refactor: GptEngduClient 프롬프트 생성 로직 전략 패턴 적용
